Apply CSS Grid properties to the body section
I applied CSS Grid properties to the body of the admin dashboard to
begin the structure and layout of the page. I added classes to the
header, sidebar, and main content and positioned them in the grid using
grid areas. To visually split the sections apart, I added background
color to each.

I also formatted all links to have no list style, set their padding and
margins to zero, and remove the underline of their anchor tags.